Added the Isolate_Paper file to the repository.
The Isolate_Paper will be a combination of the Hough Transform file and
Classify Urine file. It will peform the hough transfom on multiple images
and output a cropped image of the paper.

I am adding this file now to streamline the process of constructing a box
from the hough transform.